First steps into including logging
Logging is powered by Monolog, and the extension Monolog Cascade. The
version of the latter in use is a personal fork that removes a lot of
the clutter from the original version. To look at how Monolog Cascade is
configured, look at the example configuration files.

Right now, two loggers are always expected to exist: "main" and "error".

In future, there are plans to log a lot more, especially (for example)
memory usage when loading and processing images.