Display photo date taken from EXIF data in images
To put the date taken display between the slides and the dots of the
carousel, I had to move the dots into their own container. Fortunately,
this is fully supported by slick.

It would be nice if slick could dispatch an event just before the
current image has finished changing, rather than after, as this would
make the date taken display feel more responsive.

This still needs a bit of work:
- Gotta transition it to a proper template
- Gotta add an extended metadata tooltip/popover thing that displays
  additional metadata from images