Add metadata button to carousel display
This lets users view various bits of metadata about the photos.
Currently this is quite limited, partially due to a lack of support for
some fields in the EXIF reader package I'm using.

Also, I'm not sure that any of this code works properly for albums that
don't have the metadata permission. Will need to check that at some
point.

This commit also features a fix to the templating engine. It will now
work properly when a numeric datatype is passed in. Numeric datatypes
shouldn't pose any XSS threat, so they are put straight into the
template as-is.