Add support for displaying GPS data
This commit introduces a couple of new features.

First off is being able to show where a photo was taken based on GPS
data stored within EXIF data. Like display of image metadata, this must
be enabled on a per-path basis with the "gps" permission. When viewing
images, a map button will be displayed if that image has GPS data
available. Clicking the map button will open a new modal with a map
inside. This map is powered by Leaflet.js and Mapbox. If no Mapbox
configuration is available, no map button will be made available.

A "modal manager" has been created to manage the lifecycle of multiple
modals being open at the same time. This is based on an example I found
on codepen. One challenge I had was to ensure that the "modal-open"
class stayed on the body after closing the second modal. Currently, only
one extra modal can be open, but it shouldn't be difficult to support
arbitrary layering.