Stop vertical scrollbar appearing after downloading file in viewer modal
This was a similar issue that I had to deal with when building the modal
manager, to allow multiple modals to be open at once. In this instance,
however, trying to shoehorn the loading mask into the modal manager
would have been an unnecessarily difficult and annoying thing to do, so
I didn't do it.