Fix display issues on mobile
Since pretty much the very beginning, the site has not displayed
properly at the mobile breakpoint. Turns out this is because the main
content box, which is also the bootstrap grid container, had flex
applied to it but no width. By applying width: 100%, the container
automatically scales to fit the width of the screen for all mobile
users.

Oddly enough, the gallery Fullscreen button is still necessary because
at really small breakpoints, the gallery skin's fullscreen button
doesn't show up.