Hide gallery layer box when there is no content in it
The best way of doing this is using the :empty CSS selector, but that
requires the element to have literally no content inside of it, not even
a single space. Therefore, it is now possible to add a data attribute to
any elements within your template that you want to be trimmed
automatically.